refactor: 현재 예외 처리 구조에 맞춰 auth 패키지 리팩터링
