--- src/main/java/wooteco/subway/auth/exception/AuthorizationException.java
@@ -1,8 +1,0 @@
-package wooteco.subway.auth.exception;
-
-public class AuthorizationException extends RuntimeException {
-
-    public AuthorizationException(String message) {
-        super(String.format("인증에 실패했습니다. (%s)", message));
-    }
-}
