--- src/main/java/wooteco/subway/exception/ObjectNotFoundException.java
@@ -1,8 +1,0 @@
-package wooteco.subway.exception;
-
-public class ObjectNotFoundException extends RuntimeException {
-
-    public ObjectNotFoundException(String message) {
-        super(message);
-    }
-}
