+++ src/main/java/wooteco/subway/member/application/MemberService.java
@@ -3,7 +3,7 @@ package wooteco.subway.member.application;
 import org.springframework.dao.DataAccessException;
 import org.springframework.stereotype.Service;
 import wooteco.subway.exception.BusinessRelatedException;
-import wooteco.subway.exception.ObjectNotFoundException;
+import wooteco.subway.exception.application.ObjectNotFoundException;
 import wooteco.subway.member.dao.MemberDao;
 import wooteco.subway.member.domain.Member;
 import wooteco.subway.member.dto.MemberRequest;
